+++ src/schemaGenerator.ts
@@ -25,6 +25,7 @@ import {
   GraphQLType,
   GraphQLInterfaceType,
   GraphQLFieldMap,
+  GraphQLUnionType,
 } from 'graphql';
 
 import {
@@ -237,10 +238,13 @@ function buildSchemaFromTypeDefinitions(
 }
 
 const newExtensionDefinitionKind = 'ObjectTypeExtension';
+const interfaceExtensionDefinitionKind = 'InterfaceTypeExtension';
 
 export function extractExtensionDefinitions(ast: DocumentNode) {
   const extensionDefs = ast.definitions.filter(
-    (def: DefinitionNode) => (def.kind as any) === newExtensionDefinitionKind,
+    (def: DefinitionNode) =>
+      (def.kind as any) === newExtensionDefinitionKind ||
+      (def.kind as any) === interfaceExtensionDefinitionKind,
   );
 
   return Object.assign({}, ast, {
@@ -380,7 +384,10 @@ function addResolveFunctionsToSchema(
   resolveFunctions: IResolvers,
   resolverValidationOptions: IResolverValidationOptions = {},
 ) {
-  const { allowResolversNotInSchema = false } = resolverValidationOptions;
+  const {
+    allowResolversNotInSchema = false,
+    requireResolversForResolveType,
+  } = resolverValidationOptions;
 
   Object.keys(resolveFunctions).forEach(typeName => {
     const type = schema.getType(typeName);
@@ -397,7 +404,6 @@ function addResolveFunctionsToSchema(
     Object.keys(resolveFunctions[typeName]).forEach(fieldName => {
       if (fieldName.startsWith('__')) {
         // this is for isTypeOf and resolveType and all the other stuff.
-        // TODO require resolveType for unions and interfaces.
         type[fieldName.substring(2)] = resolveFunctions[typeName][fieldName];
         return;
       }
@@ -453,6 +459,29 @@ function addResolveFunctionsToSchema(
         setFieldProperties(field, fieldResolve);
       }
     });
+  });
+
+  checkForResolveTypeResolver(schema, requireResolversForResolveType);
+}
+
+// If we have any union or interface types throw if no there is no resolveType or isTypeOf resolvers
+function checkForResolveTypeResolver(schema: GraphQLSchema, requireResolversForResolveType?: boolean) {
+  Object.keys(schema.getTypeMap())
+  .map(typeName => schema.getType(typeName))
+  .forEach((type: GraphQLUnionType | GraphQLInterfaceType) => {
+    if (!(type instanceof GraphQLUnionType || type instanceof GraphQLInterfaceType)) {
+      return;
+    }
+    if (!type.resolveType) {
+      if (requireResolversForResolveType === false) {
+        return;
+      }
+      if (requireResolversForResolveType === true) {
+        throw new SchemaError(`Type "${type.name}" is missing a "resolveType" resolver`);
+      }
+      // tslint:disable-next-line:max-line-length
+      console.warn(`Type "${type.name}" is missing a "resolveType" resolver. Pass false into "resolverValidationOptions.requireResolversForResolveType" to disable this warning.`);
+    }
   });
 }